fix fa in ascend_speed by torch_npu
